chore: 챌린지 이미지 업로드 오류 수정
chore: 챌린지 이미지 업로드 오류 수정
