--- src/main/java/com/mouse/challenge/domain/ChallengeGroup.java
@@ -1,27 +1,0 @@
-package com.mouse.challenge.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import java.util.List;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class ChallengeGroup {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "challenge_group_id")
-    private Long id;
-
-    private String name;
-
-    @OneToMany(mappedBy = "challengeGroup")
-    private List<ChallengePost> challengePost;
-}
