+++ src/main/java/com/mouse/challenge/dto/ChallengePostUpdateRequest.java
@@ -1,10 +1,8 @@
 package com.mouse.challenge.dto;
 
-import com.mouse.challenge.domain.Image;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -13,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 @NoArgsConstructor
 public class ChallengePostUpdateRequest {
     private Long postId;
-    private Long userId;
+    private String userId;
     private Long challengeGroupId;
     private String content;
     private List<MultipartFile> images;
