+++ src/main/java/com/mouse/challenge/entity/ChallengeGroup.java
@@ -1,26 +1,27 @@
-package com.mouse.challenge.domain;
+package com.mouse.challenge.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class User {
+public class ChallengeGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "challenge_group_id")
     private Long id;
 
-    @Column(nullable = false, unique = true)
-    private String username;
+    private String name;
 
-    @Column(nullable = false)
-    private String password;
-
+    @OneToMany(mappedBy = "challengeGroup")
+    private List<ChallengePost> challengePost;
 }
