+++ src/main/java/com/mouse/challenge/entity/ChallengePost.java
@@ -1,4 +1,4 @@
-package com.mouse.challenge.domain;
+package com.mouse.challenge.entity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -10,7 +10,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
-import jakarta.persistence.PrePersist;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Data;
