+++ src/main/java/com/mouse/challenge/repository/ChallengePostRepository.java
@@ -1,11 +1,10 @@
 package com.mouse.challenge.repository;
 
-import com.mouse.challenge.domain.ChallengePost;
-import jakarta.persistence.EntityManager;
+import com.mouse.challenge.entity.ChallengePost;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -16,7 +15,7 @@ public interface ChallengePostRepository extends JpaRepository<ChallengePost, Lo
         + " join fetch p.images i"
         + " join fetch p.user u"
         + " where p.id = :postId")
-    ChallengePost findByIdFetchJoin(Long postId);
+    ChallengePost findByIdFetchJoin(@Param("postId") Long postId);
 
     @Query("select p from ChallengePost p"
         + " join fetch p.challengeGroup g"
